tdf#114260: Add gear button to Customize dialog
  * Allow renaming custom toolbars, custom menus, and standard sub-menus.
  * Move plus and minus buttons into the gear button

+++ cui/source/customize/SvxMenuConfigPage.cxx
@@ -117,10 +117,8 @@ SvxMenuConfigPage::SvxMenuConfigPage(vcl::Window *pParent, const SfxItemSet& rSe
     m_pContentsListBox->SetSelectHdl(
         LINK( this, SvxMenuConfigPage, SelectMenuEntry ) );
 
-    m_pPlusBtn->SetClickHdl(
-        LINK( this, SvxMenuConfigPage, AddMenuHdl ) );
-    m_pMinusBtn->SetClickHdl(
-        LINK( this, SvxMenuConfigPage, RemoveMenuHdl ) );
+    m_pGearBtn->SetSelectHdl(
+        LINK( this, SvxMenuConfigPage, GearHdl ) );
 
     m_pCommandCategoryListBox->SetSelectHdl(
         LINK( this, SvxMenuConfigPage, SelectCategory ) );
@@ -147,18 +145,9 @@ SvxMenuConfigPage::SvxMenuConfigPage(vcl::Window *pParent, const SfxItemSet& rSe
 
     if ( !bIsMenuBar )
     {
-        // Context menus cannot be added/removed
-        // These height requests are needed to keep the height of
-        // the top level list boxes consistent for all tabs
-        m_pTopLevelListBox->set_height_request(
-            m_pPlusBtn->GetOptimalSize().Height() );
-        m_pCommandCategoryListBox->set_height_request(
-            m_pPlusBtn->GetOptimalSize().Height() );
-
-        m_pPlusBtn->Disable();
-        m_pMinusBtn->Disable();
-        m_pPlusBtn->Hide();
-        m_pMinusBtn->Hide();
+        //TODO: Remove this when the gear button is implemented for context menus
+        m_pGearBtn->Disable();
+        m_pGearBtn->Hide();
     }
     else
     {
@@ -225,6 +214,16 @@ void SvxMenuConfigPage::UpdateButtonStates()
     m_pRemoveCommandButton->Enable( bIsValidSelection );
 
     m_pModifyBtn->Enable( bIsValidSelection && !bIsSeparator);
+
+    //Handle the gear button
+    if (m_bIsMenuBar)
+    {
+        SvxConfigEntry* pMenuData = GetTopLevelSelection();
+        PopupMenu* pGearPopup = m_pGearBtn->GetPopupMenu();
+        // Add option (gear_add) will always be enabled
+        pGearPopup->EnableItem( "gear_delete", pMenuData->IsDeletable() );
+        pGearPopup->EnableItem( "gear_rename", pMenuData->IsRenamable() );
+    }
 }
 
 void SvxMenuConfigPage::DeleteSelectedTopLevel()
@@ -298,9 +297,6 @@ IMPL_LINK_NOARG( SvxMenuConfigPage, SelectMenu, ListBox&, void )
 
     if ( pMenuData )
     {
-        // Built-in menus cannot be deleted
-        m_pMinusBtn->Enable( pMenuData->IsDeletable() );
-
         SvxEntries* pEntries = pMenuData->GetEntries();
         SvxEntries::const_iterator iter = pEntries->begin();
 
@@ -314,22 +310,60 @@ IMPL_LINK_NOARG( SvxMenuConfigPage, SelectMenu, ListBox&, void )
     UpdateButtonStates();
 }
 
-IMPL_LINK_NOARG( SvxMenuConfigPage, AddMenuHdl, Button *, void )
+IMPL_LINK( SvxMenuConfigPage, GearHdl, MenuButton *, pButton, void )
 {
-    VclPtrInstance<SvxMainMenuOrganizerDialog> pDialog(
-        nullptr, GetSaveInData()->GetEntries(), nullptr, true );
+    OString sIdent = pButton->GetCurItemIdent();
 
-    if ( pDialog->Execute() == RET_OK )
+    if (sIdent == "gear_add")
     {
-        GetSaveInData()->SetEntries( pDialog->GetEntries() );
-        ReloadTopLevelListBox( pDialog->GetSelectedEntry() );
-        GetSaveInData()->SetModified();
-    }
-}
+        VclPtrInstance<SvxMainMenuOrganizerDialog> pDialog(
+            nullptr, GetSaveInData()->GetEntries(), nullptr, true );
 
-IMPL_LINK_NOARG( SvxMenuConfigPage, RemoveMenuHdl, Button *, void )
-{
-    DeleteSelectedTopLevel();
+        if ( pDialog->Execute() == RET_OK )
+        {
+            GetSaveInData()->SetEntries( pDialog->GetEntries() );
+            ReloadTopLevelListBox( pDialog->GetSelectedEntry() );
+            GetSaveInData()->SetModified();
+        }
+    }
+    else if (sIdent == "gear_delete")
+    {
+        DeleteSelectedTopLevel();
+    }
+    else if (sIdent == "gear_rename")
+    {
+        SvxConfigEntry* pMenuData = GetTopLevelSelection();
+
+        OUString sCurrentName( SvxConfigPageHelper::stripHotKey( pMenuData->GetName() ) );
+        OUString sDesc = CuiResId( RID_SVXSTR_LABEL_NEW_NAME );
+
+        VclPtrInstance< SvxNameDialog > pNameDialog( this, sCurrentName, sDesc );
+        pNameDialog->SetHelpId(HID_SVX_CONFIG_RENAME_MENU);
+        pNameDialog->SetText( CuiResId( RID_SVXSTR_RENAME_MENU ) );
+
+        if ( pNameDialog->Execute() == RET_OK )
+        {
+            OUString sNewName;
+            pNameDialog->GetName( sNewName );
+
+            if ( sCurrentName == sNewName )
+                return;
+
+            pMenuData->SetName( sNewName );
+
+            ReloadTopLevelListBox();
+
+            GetSaveInData()->SetModified();
+        }
+    }
+    else
+    {
+        //This block should never be reached
+        SAL_WARN("cui.customize", "Unknown gear menu option: " << sIdent);
+        return;
+    }
+
+    UpdateButtonStates();
 }
 
 IMPL_LINK_NOARG( SvxMenuConfigPage, SelectCategory, ListBox&, void )
